Refactored code, split up debouncer yet again, "debouncer_state.cpp" is now the only part that needs testing

+++ main/src/persistent_store.cpp
@@ -2,14 +2,13 @@
 #include <sys/stat.h>
 #include <sys/unistd.h>
 
+#include "sdkconfig.h"
 #include "esp_log.h"
 #include "esp_spiffs.h"
 
 #include "persistent_store.hpp"
 
 #define RESET true
-
-static const char* TAG = "[PersistentStore]";
 
 size_t spiffsFreeSpace(const char* partition_label) {
     size_t used = 0;
@@ -19,8 +18,10 @@ size_t spiffsFreeSpace(const char* partition_label) {
     return total - used;
 }
 
-PersistentStore::PersistentStore(): m_can_recover(false) {
-    ESP_LOGD(TAG, "Persistent store ctor called");
+PersistentStore::PersistentStore(bool reset):   m_can_recover(false),
+                                                m_reset(reset)
+{
+    ESP_LOGD("PersistentStore", "Persistent store ctor called");
 
     m_spiffs_conf.base_path = "/spiffs";
     m_spiffs_conf.max_files = 1;
@@ -33,11 +34,11 @@ PersistentStore::PersistentStore(): m_can_recover(false) {
     size_t used = 0;
     ESP_ERROR_CHECK(esp_spiffs_info(m_spiffs_conf.partition_label, &total, &used));
 
-    ESP_LOGD(TAG, "Partition size: Total -> %d, Used -> %d", total, used);
+    ESP_LOGD("PersistentStore", "Partition size: Total -> %d, Used -> %d", total, used);
 
-    #if RESET == true
+    if (m_reset) {
         unlink("/spiffs/log");
-    #endif
+    }
 
     struct stat st;
     if (stat("/spiffs/log", &st) == 0) {
@@ -45,10 +46,10 @@ PersistentStore::PersistentStore(): m_can_recover(false) {
     }
 
     const char* flags = m_can_recover ? "a+b" : "w+b";
-    ESP_LOGD(TAG, "Can recover: %s. Setting flags: %s", m_can_recover ? "yes": "no", flags);
+    ESP_LOGD("PersistentStore", "Can recover: %s. Setting flags: %s", m_can_recover ? "yes": "no", flags);
     FILE* file = fopen("/spiffs/log", flags);
     if (file == NULL) {
-        ESP_LOGE(TAG, "Failed to open file");
+        ESP_LOGE("PersistentStore", "Failed to open file");
     }
 
     m_log_file = file;
@@ -70,43 +71,45 @@ void PersistentStore::addPass(Pass* pass) {
     }
 }
 
-void PersistentStore::printLog() {
-    // Save initial position
-    fpos_t initial_position;
-    fgetpos(m_log_file, &initial_position);
+#if CONFIG_LOG_DEFAULT_LEVEL == 4
+    void PersistentStore::printLog() {
+        // Save initial position
+        fpos_t initial_position;
+        fgetpos(m_log_file, &initial_position);
 
-    ESP_LOGD(TAG, "Current position: %ld", initial_position);
+        ESP_LOGD("PersistentStore", "Current position: %ld", initial_position);
 
-    size_t length = ftell(m_log_file);
+        size_t length = ftell(m_log_file);
 
-    if (length == 0) {
-        ESP_LOGD(TAG, "Log file is empty, aborting print");
-        return;
+        if (length == 0) {
+            ESP_LOGD("PersistentStore", "Log file is empty, aborting print");
+            return;
+        }
+
+        ESP_LOGD("PersistentStore", "File length: %u", length);
+        ESP_LOGD("PersistentStore", "Limits: %u", length);
+        ESP_LOGD("PersistentStore", "Increment: %u", (sizeof(int64_t) + sizeof(uint16_t)));
+
+        // Print all entries between start and initial position
+        for (auto i = 0; i <= length; i += (sizeof(int64_t) + sizeof(uint16_t))) {
+            int64_t time;
+            uint16_t duration;
+
+            // Jump to start of pair
+            fseek(m_log_file, i, SEEK_SET);
+            // Read time from log file
+            auto time_read = fread(&time, sizeof(time), 1, m_log_file);
+
+            // Jump to duration
+            fseek(m_log_file, i + 8, SEEK_SET);
+            
+            // Read duration
+            fread(&duration, sizeof(duration), 1, m_log_file);
+
+            ESP_LOGD("PersistentStore", "Pass: Time -> %lld, Duration -> %d\n", time, duration);
+        }
+
+        // Restore initial position
+        fsetpos(m_log_file, &initial_position);
     }
-
-    ESP_LOGD(TAG, "File length: %u", length);
-    ESP_LOGD(TAG, "Limits: %u", length);
-    ESP_LOGD(TAG, "Increment: %u", (sizeof(int64_t) + sizeof(uint16_t)));
-
-    // Print all entries between start and initial position
-    for (auto i = 0; i <= length; i += (sizeof(int64_t) + sizeof(uint16_t))) {
-        int64_t time;
-        uint16_t duration;
-
-        // Jump to start of pair
-        fseek(m_log_file, i, SEEK_SET);
-        // Read time from log file
-        fread(&time, sizeof(time), 1, m_log_file);
-
-        // Jump to duration
-        fseek(m_log_file, i + 8, SEEK_SET);
-        // Read duration
-
-        fread(&duration, sizeof(duration), 1, m_log_file);
-
-        printf("Pass: Time -> %lld, Duration -> %d\n", time, duration);
-    }
-
-    // Restore initial position
-    fsetpos(m_log_file, &initial_position);
-}
+#endif
